requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
python-docx==1.1.0
openpyxl==3.1.2
google-genai==1.21.1
python-dotenv==1.0.0
aiofiles==23.2.1
scipy==1.11.4
